missing list products, user_id from header and swagger

+++ internal/endpoint/objects.go
@@ -59,8 +59,8 @@ type (
 
 	ListProductsByCategoryRequest struct {
 		ID     string `json:"id"`
-		Limit  int    `json:"limit"`
-		Offset int    `json:"offset"`
+		Limit  int64  `json:"limit"`
+		Offset int64  `json:"offset"`
 	}
 
 	DeleteProductRequest struct {
